Refactor require path
