+++ tests/test-protocol-error.js
@@ -1,5 +1,5 @@
 let assert = require('assert');
-let Thrift = require('../thrift');
+let Thrift = require('../');
 
 it('protocol error must be caught', done => {
