+++ tests/test-type-conversion.js
@@ -1,5 +1,5 @@
 let assert = require('assert');
-let Thrift = require('../thrift');
+let Thrift = require('../');
 
 it('type conversion must be supported', done => {
